pyproject.toml
[project]
name = "streamlit-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.0",
    "langchain-aws>=1.1.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
